+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Each change log are assumed to be accompanied by associated unit test and documentation updates.
 
+## master
+- added ability to set custom setValue() method for auto complete (#7)
+- added ability to define custom getData() method for auto complete (#6)
+- fixed issue with new indicator not work whe initial search for data returns nothing (#4)
+- all tests now exectue against both input and textarea elements (#3)
+- refactored tests to be more maintainable
+
 ## 0.5.0
 
 - added ability to be able to put free form new indicator in with the list of options